Update Implementation.py
Fixed a few small details (ex: if self.data[i] == None -> if not self.data[i] on line 29).

Also changed the keys method for if there is collision to be slightly more efficient. 

Previous: Did not check to see if the current bucket had more than one list or not and just looped over the entire bucket each time.
Added: Added a check to see if current bucket has more than one key/value pair. This prevents it from needing to do that extra for loop for just one value.